@@ -1,14 +1,8 @@
 import { Node } from "estree-walker";
-import type { parse } from "svelte/compiler";
+import type { AST } from "svelte/compiler";
 
-function parseModern(source: string) {
-  return parse(source, { modern: true });
-}
-
-type Root = ReturnType<typeof parseModern>;
-
-export type Script = NonNullable<Root["instance"]>;
-export type Fragment = Root["fragment"];
+export type Script = NonNullable<AST.Root["instance"]>;
+export type Fragment = AST.Root["fragment"];
 
 export interface ClassesDefinition {
   start: number;
